Refactors services
