+++ src/app/best-deal/best-deal.component.ts
@@ -2,9 +2,9 @@ import { Component } from '@angular/core';
 import { FormBuilder, FormGroup } from '@angular/forms';
 import { Subject, takeUntil } from 'rxjs';
 import { Item, Store } from '../shared/model/common.model';
-import { ItemService } from '../service/item/item.service';
+import { ItemService } from '../service/item.service';
 import { APP } from '../shared/constant/app.const';
-import { SessionDataService } from '../service/session/session-data.service';
+import { SessionDataService } from '../service/session-data.service';
 import { CartItemType, CartType, SessionData } from '../shared/model/SessionData.model';
 import { Router } from '@angular/router';
