--- src/app/service/category/category.service.spec.ts
@@ -1,16 +1,0 @@
-// import { TestBed } from '@angular/core/testing';
-
-// import { CategoryService } from './category.service';
-
-// describe('CategoryService', () => {
-//   let service: CategoryService;
-
-//   beforeEach(() => {
-//     TestBed.configureTestingModule({});
-//     service = TestBed.inject(CategoryService);
-//   });
-
-//   it('should be created', () => {
-//     expect(service).toBeTruthy();
-//   });
-// });
