+++ src/app/service/log.service.spec.ts
